aider: Added support for aborting long-running operations by introducing an abort channel and using tokio::select! to listen for abort signals.
# Aider chat conversation:

USER: Hello. Please note the loop in run(). I need a channel or some way that I can send in a message to abort whatever the current command or message input it is processing. For example the `send_request` can take a long time, or the `call_function` also might need to be aborted. But I want it to continue looping on input. I plan to use tokio select when processing the long request etc. But I think I need to pass in another channel or something into run()

ASSISTANT: To achieve this, we can introduce a new channel that will be used to send abort signals. We will use `tokio::select!` to listen on both the user input channel and the abort channel simultaneously. If an abort message is received, we will break the current operation and continue with the next iteration of the loop.

Here are the steps:

1. Add a new `abort_receiver` field to the `Agent` struct.
2. Modify the `new` function to accept an additional `abort_receiver` parameter and initialize the new field.
3. Modify the `run` function to use `tokio::select!` to listen on both the user input channel and the abort channel.

Here are the *SEARCH/REPLACE* blocks:

src/agent/mod.rs
```rust
<<<<<<< SEARCH
pub struct Agent {
    username: String,
    functions: Vec<ChatCompletionFunctions>,
    session_id: usize,
    log: ChatLog,
    model: String,
    chat: OpenAIChat,
    handler: Handler,
    receiver: flume::Receiver<String>,
    reply_sender: flume::Sender<ChatUIMessage>,
}
=======
pub struct Agent {
    username: String,
    functions: Vec<ChatCompletionFunctions>,
    session_id: usize,
    log: ChatLog,
    model: String,
    chat: OpenAIChat,
    handler: Handler,
    receiver: flume::Receiver<String>,
    reply_sender: flume::Sender
^C KeyboardInterrupt

USER: hm. the thing is, I need the `send_request` to internally be able to receive an abort signal somehow. so if I am only checking for a halt command on the outer loop but not monitoring inside the request sender (which streams) or the `call_function` then it won't work. So I need some mechanism that will allow that without waiting for those two things inside the loop to complete (maybe using tokio spawn or just passing the internal abort channel into those calls or something)

ASSISTANT: In that case, you can pass the abort receiver into the `send_request` and `call_function` methods. Inside these methods, you can use `tokio::select!` to listen for abort signals while processing the request or function call. If an abort signal is received, you can immediately return from the method.

Here are the steps:

1. Modify the `send_request` and `call_function` method signatures to accept an additional `abort_receiver` parameter.
2. Inside these methods, use `tokio::select!` to listen for abort signals while processing the request or function call.
3. If an abort signal is received, immediately return from the method.

Here are the *SEARCH/REPLACE* blocks:

src/agent/mod.rs
```rust
<<<<<<< SEARCH
    pub fn call(&mut self, fn_name: &str, args_json: &str) -> Result<String> {
        call_function(&mut self.handler, fn_name, args_json)
    }
=======
    pub fn call(&mut self, fn_name: &str, args_json: &str, abort_receiver: &flume::Receiver<()>) -> Result<String> {
        tokio::select! {
            result = call_function(&mut self.handler, fn_name, args_json) => result,
            _ = abort_receiver.recv_async() => Err(anyhow!("Aborted")),
        }
    }
>>>>>>> REPLACE
```

src/agent/mod.rs
```rust
<<<<<<< SEARCH
    pub async fn process_fn_call(&mut self, fn_name: &str, fn_args: &str) -> Result<()> {
        self.log
            .add(fn_call_msg(&fn_name.to_string(), &fn_args.to_string())?);

        let output = self.call(fn_name, fn_args)?;
        self.log
            .add(fn_result_msg(&s!(fn_name), &output.to_string())?);
        println!("Trying to send func call back");
        self.reply_sender
            .send_async(ChatUIMessage::FunctionCall {
                name: s!(fn_name),
                params: s!(fn_args),
                result: output,
            })
            .await?;
        println!("Function call: {}({})", fn_name, fn_args);

        let next_step = self.call_ret_string("evalExitStage", "{}")?;
        if next_step.contains("Function not found") {
        } else {
            if next_step != "" && next_step != "()" {
                self.next_stage(&next_step).await?;
            }
        };

        Ok(())
    }
=======
    pub async fn process_fn_call(&mut self, fn_name: &str, fn_args: &str, abort_receiver: &flume::Receiver<()>) -> Result<()> {
        self.log
            .add(fn_call_msg(&fn_name.to_string(), &fn_args.to_string())?);

        let output = self.call(fn_name, fn_args, abort_receiver)?;
        self.log
            .add(fn_result_msg(&s!(fn_name), &output.to_string())?);
        println!("Trying to send func call back");
        self.reply_sender
            .send_async(ChatUIMessage::FunctionCall {
                name: s!(fn_name),
                params: s!(fn_args),
                result: output,
            })
            .await?;
        println!("Function call: {}({})", fn_name, fn_args);

        let next_step = self.call_ret_string("evalExitStage", "{}")?;
        if next_step.contains("Function not found") {
        } else {
            if next_step != "" && next_step != "()" {
                self.next_stage(&next_step).await?;
            }
        };

        Ok(())
    }
>>>>>>> REPLACE
```

Please note that you will need to modify the `call_function` and `send_request` functions to accept and handle the abort receiver.